feat(model-server): add dedicated metric handler (MODELIX-720)
