+++ model-server/src/main/kotlin/org/modelix/model/server/Main.kt
@@ -24,7 +24,6 @@ import io.ktor.server.engine.embeddedServer
 import io.ktor.server.html.respondHtmlTemplate
 import io.ktor.server.http.content.resources
 import io.ktor.server.http.content.static
-import io.ktor.server.metrics.micrometer.MicrometerMetrics
 import io.ktor.server.netty.Netty
 import io.ktor.server.netty.NettyApplicationEngine
 import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
@@ -32,7 +31,6 @@ import io.ktor.server.plugins.cors.routing.CORS
 import io.ktor.server.plugins.forwardedheaders.ForwardedHeaders
 import io.ktor.server.plugins.swagger.swaggerUI
 import io.ktor.server.resources.Resources
-import io.ktor.server.response.respond
 import io.ktor.server.response.respondText
 import io.ktor.server.routing.IgnoreTrailingSlash
 import io.ktor.server.routing.Routing
@@ -41,16 +39,6 @@ import io.ktor.server.routing.routing
 import io.ktor.server.websocket.WebSockets
 import io.ktor.server.websocket.pingPeriod
 import io.ktor.server.websocket.timeout
-import io.micrometer.core.instrument.binder.jvm.ClassLoaderMetrics
-import io.micrometer.core.instrument.binder.jvm.JvmGcMetrics
-import io.micrometer.core.instrument.binder.jvm.JvmMemoryMetrics
-import io.micrometer.core.instrument.binder.jvm.JvmThreadMetrics
-import io.micrometer.core.instrument.binder.system.FileDescriptorMetrics
-import io.micrometer.core.instrument.binder.system.ProcessorMetrics
-import io.micrometer.core.instrument.binder.system.UptimeMetrics
-import io.micrometer.core.instrument.distribution.DistributionStatisticConfig
-import io.micrometer.prometheus.PrometheusConfig
-import io.micrometer.prometheus.PrometheusMeterRegistry
 import kotlinx.html.a
 import kotlinx.html.h1
 import kotlinx.html.li
@@ -65,6 +53,7 @@ import org.modelix.model.server.handlers.DeprecatedLightModelServer
 import org.modelix.model.server.handlers.HistoryHandler
 import org.modelix.model.server.handlers.KeyValueLikeModelServer
 import org.modelix.model.server.handlers.ModelReplicationServer
+import org.modelix.model.server.handlers.MetricsHandler
 import org.modelix.model.server.handlers.RepositoriesManager
 import org.modelix.model.server.handlers.RepositoryOverview
 import org.modelix.model.server.store.IStoreClient
@@ -173,7 +162,7 @@ object Main {
             val historyHandler = HistoryHandler(localModelClient, repositoriesManager)
             val contentExplorer = ContentExplorer(localModelClient, repositoriesManager)
             val modelReplicationServer = ModelReplicationServer(repositoriesManager)
-            val appMicrometerRegistry = PrometheusMeterRegistry(PrometheusConfig.DEFAULT)
+            val metricsHandler = MetricsHandler()
             val ktorServer: NettyApplicationEngine = embeddedServer(Netty, port = port) {
                 install(Routing)
                 installAuthentication(unitTestMode = !KeycloakUtils.isEnabled())
@@ -198,26 +187,6 @@ object Main {
                     allowMethod(HttpMethod.Put)
                     allowMethod(HttpMethod.Post)
                 }
-                install(MicrometerMetrics) {
-                    registry = appMicrometerRegistry
-                    distributionStatisticConfig = DistributionStatisticConfig.Builder()
-                        .percentilesHistogram(true)
-                        .maximumExpectedValue(Duration.ofSeconds(20).toNanos().toDouble())
-                        .serviceLevelObjectives(
-                            Duration.ofMillis(100).toNanos().toDouble(),
-                            Duration.ofMillis(500).toNanos().toDouble(),
-                        )
-                        .build()
-                    meterBinders = listOf(
-                        ClassLoaderMetrics(),
-                        JvmMemoryMetrics(),
-                        JvmGcMetrics(),
-                        ProcessorMetrics(),
-                        JvmThreadMetrics(),
-                        FileDescriptorMetrics(),
-                        UptimeMetrics(),
-                    )
-                }
 
                 modelServer.init(this)
                 historyHandler.init(this)
@@ -225,12 +194,9 @@ object Main {
                 contentExplorer.init(this)
                 jsonModelServer.init(this)
                 modelReplicationServer.init(this)
+                metricsHandler.init(this)
                 routing {
-                    get("/metrics") {
-                        call.respond(appMicrometerRegistry.scrape())
-                    }
-
-                    static("/public") {
+                     static("/public") {
                         resources("public")
                     }
                     get("/") {
